Unify app token authentication between Darc and PCS (#3649)
