+++ src/Maestro/Client/src/MaestroApi.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Azure.Core;
 using Newtonsoft.Json.Linq;
 using System;
 using System.IO;
@@ -25,14 +24,6 @@ namespace Microsoft.DotNet.Maestro.Client
 
     public partial class MaestroApi
     {
-        public const string ProductionBuildAssetRegistryBaseUri = "https://maestro.dot.net/";
-
-        public const string StagingBuildAssetRegistryBaseUri = "https://maestro.int-dot.net/";
-
-        public const string OldProductionBuildAssetRegistryBaseUri = "https://maestro-prod.westus2.cloudapp.azure.com/";
-
-        public const string OldStagingBuildAssetRegistryBaseUri = "https://maestro-int.westus2.cloudapp.azure.com/";
-
         // Special error handler to consumes the generated MaestroApi code. If this method returns without throwing a specific exception
         // then a generic RestApiException is thrown.
         partial void HandleFailedRequest(RestApiException ex)
@@ -59,52 +50,6 @@ namespace Microsoft.DotNet.Maestro.Client
                 throw new AuthenticationException("Unauthorized access while trying to access Maestro API. " +
                     "Please make sure the PAT you're using is valid.");
             }
-        }
-
-        /// <summary>
-        /// Creates a credential based on parameters provided.
-        /// </summary>
-        /// <param name="barApiBaseUri">BAR API URI used to determine the right set of credentials (INT vs PROD)</param>
-        /// <param name="disableInteractiveAuth">Whether to include interactive login flows</param>
-        /// <param name="barApiToken">Token to use for the call. If none supplied, will try other flows.</param>
-        /// <param name="federatedToken">Federated token to use for the call. If none supplied, will try other flows.</param>
-        /// <param name="managedIdentityId">Managed Identity to use for the auth</param>
-        /// <returns>Credential that can be used to call the Maestro API</returns>
-        public static TokenCredential CreateApiCredential(
-            string barApiBaseUri,
-            bool disableInteractiveAuth,
-            string? barApiToken = null,
-            string? federatedToken = null,
-            string? managedIdentityId = null)
-        {
-            // 1. BAR or Entra token that can directly be used to authenticate against Maestro
-            if (!string.IsNullOrEmpty(barApiToken))
-            {
-                return new MaestroApiTokenCredential(barApiToken!);
-            }
-
-            barApiBaseUri ??= ProductionBuildAssetRegistryBaseUri;
-
-            // 2. Federated token that can be used to fetch an app token (for CI scenarios)
-            if (!string.IsNullOrEmpty(federatedToken))
-            {
-                return MaestroApiCredential.CreateFederatedCredential(barApiBaseUri, federatedToken!);
-            }
-
-            // 3. Managed identity (for server-to-server scenarios - e.g. PCS->Maestro)
-            if (!string.IsNullOrEmpty(managedIdentityId))
-            {
-                return MaestroApiCredential.CreateManagedIdentityCredential(barApiBaseUri, managedIdentityId!);
-            }
-
-            // 4. Azure CLI authentication setup by the caller (for CI scenarios)
-            if (disableInteractiveAuth)
-            {
-                return MaestroApiCredential.CreateNonUserCredential(barApiBaseUri);
-            }
-
-            // 5. Interactive login (user-based scenario)
-            return MaestroApiCredential.CreateUserCredential(barApiBaseUri);
         }
     }
 }
